% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinBatcher-package.R
\docType{package}
\name{ProteinBatcher-package}
\alias{ProteinBatcher}
\alias{ProteinBatcher-package}
\title{ProteinBatcher: Batch Processing and Harmonization of Quantitative Proteomics Table}
\description{
ProteinBatcher provides utilities to read, validate, and harmonize quantitative proteomics result tables generated by DIA-NN from data-independent acquisition mass spectrometry experiments. The package enables batch processing of multiple DIA-NN output files, consistent handling of protein identifiers and associated metadata, and transformation of tabular data into analysis-ready formats. It facilitates downstream differential abundance analysis using linear modeling frameworks based on limma, with explicit support for multivariable experimental designs including batch effects and other confounding covariates, as well as correlation-aware designs via duplicateCorrelation.
}
\author{
\strong{Maintainer}: Aitor Moruno-Cuenca \email{morunoaitor@gmail.com} (\href{https://orcid.org/0009-0009-8133-2552}{ORCID})

Other contributors:
\itemize{
  \item Dr. Sergi Sayols-Puig \email{sergi.sayolspuig@almirall.com} (\href{https://orcid.org/0000-0002-3877-4170}{ORCID}) [contributor]
  \item Dr. Bruna Oriol-Tordera \email{bruna.orioltordera@almirall.com} (\href{https://orcid.org/0000-0002-2714-9097}{ORCID}) [contributor]
  \item Dr. Hiba Salim \email{hiba.salim@crg.eu} (\href{https://orcid.org/0000-0002-2168-0258}{ORCID}) [data contributor]
  \item Dr. Francesc Fernández-Albert \email{francesc.fernandez@almirall.com} (\href{https://orcid.org/0000-0001-5561-0701}{ORCID}) [thesis advisor]
  \item Prof. Dr. Alexandre Perera-Lluna \email{alexandre.perera@upc.edu} (\href{https://orcid.org/0000-0001-6427-851X}{ORCID}) [thesis advisor]
  \item Dr. Guadalupe Espadas-García \email{guadalupe.espadas@crg.eu} (\href{https://orcid.org/0000-0002-6415-8013}{ORCID}) [contributor, data contributor]
  \item Dr. Elisa Monzón-Casanova \email{elisa.monzoncasanova@almirall.com} (\href{https://orcid.org/0000-0001-6617-6138}{ORCID}) [contributor, funder]
  \item Prof. Dr. Eduard Sabidó-Aguade \email{eduard.sabido@upf.edu} (\href{https://orcid.org/0000-0001-6506-7714}{ORCID}) [researcher, contributor, data contributor, funder]
}

}
\keyword{internal}
